refactor, feat: 라우터 변경 및 추가
1. 라우터 users.js -> auth.js로 변경
2. 유저에 관한 라우터 user.js와 기사에 관한 라우터 text.js 추가
3. 유저페이지 userPage.pug 생성
