+++ news-site/app.js
@@ -1,19 +1,21 @@
-var createError = require('http-errors');
-var express = require('express');
-var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
-var session = require('express-session');
-var flash = require('connect-flash');
+const createError = require('http-errors');
+const express = require('express');
+const path = require('path');
+const cookieParser = require('cookie-parser');
+const logger = require('morgan');
+const session = require('express-session');
+const flash = require('connect-flash');
 const passport = require('passport');
 const passportConfig = require('./passport/index');
 const connect = require('./schemas/index');
 require('dotenv').config();
 
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
+const indexRouter = require('./routes/index');
+const authRouter = require('./routes/auth');
+const userRouter = require('./routes/user');
+const textRouter = require('./routes/text');
 
-var app = express();
+const app = express();
 connect();
 passportConfig(passport);
 
@@ -39,7 +41,9 @@ app.use(passport.initialize());
 app.use(passport.session());
 
 app.use('/', indexRouter);
-app.use('/users', usersRouter);
+app.use('/auth', authRouter);
+app.use('/user', userRouter);
+app.use('/text', textRouter);
 
 app.use(function (req, res, next) {
   next(createError(404));
