+++ news-site/views/login.pug
@@ -1,7 +1,7 @@
 extends sign
 
 block content
-    form(id="loginContainer" action='/users/login', method='post')
+    form(id="loginContainer" action='/auth/login', method='post')
         input(type="text" name="inputID" id="inputID" placeholder="아이디")
         p(id="noteID")
         input(type="password" name="inputPwd" id="inputPwd" placeholder="비밀번호")
@@ -10,5 +10,5 @@ block content
         p(id="noteLogin") #{note}
         a(id='signUpBtn' href='http://localhost:3000/signup' role='button')
             div 회원가입
-    form(action='/users/google', method='get')
+    form(action='/auth/google', method='get')
         input(type="submit" value="google")
