+++ news-site/views/userPage.pug
@@ -12,3 +12,6 @@ html
             div 로그아웃
         div(id='areaLogo')    
           h1 Circus News
+    article
+        a(herf='http://localhost:3000/text')
+            div 기사쓰기
